Split functions out of Main, created CheckIn method, expanded Room class, Ryker laid groundwork for future functionality
