+++ HotelApp.css
@@ -1,0 +1,9 @@
+#textWindow {
+    width: 400px;
+    height:300px;
+    border: 2px solid black;
+}
+
+#textInput {
+    width: 400px;
+}
