+++ HotelApp.html
@@ -1,7 +1,14 @@
 <html>
     <head>
-        <script src="HotelApp.js"></script>
+        <!--First script loads jquery from google, second script is our javascript file-->
+        <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.1/jquery.min.js" type="text/javascript"></script>
+        <script src="./HotelApp.js" type="module"></script>
+        <link rel="stylesheet" href="./HotelApp.css"></link>
     </head>
     <body>
-    </body>
-</html>
+        <div id="textWindow"> TODO</div>
+        <input type="text" id="textInput" placeholder="User Input"></input>
+        <br/>
+        <button id="startMain">Start Hotel App</button>
+    </body> 
+</html>
